requirements.txt
# Core Framework
Flask==3.0.0
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-auth==2.0.0
plotly==5.18.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.2

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Database
sqlalchemy==2.0.23

# Scheduling
APScheduler==3.10.4

# Email
python-dotenv==1.0.0

# Utilities
pytz==2023.3
